Various statistics fixes, first foray into having auth working for SSL requests.

+++ daemon/Libraries/Core/Statistics/IStatistician.cs
@@ -4,6 +4,6 @@ namespace Spectero.daemon.Libraries.Core.Statistics
 {
     public interface IStatistician
     {
-        Task<bool> Update<T> (double bytes, DataFlowDirections direction) where T : new();
+        Task<bool> Update<T> (long bytes, DataFlowDirections direction) where T : new();
     }
 }
